+++ tests/Silverback.Integration.Tests.E2E/TestTypes/Messages/TestEventWithHeaders.cs
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Sergio Aquilini
 // This code is licensed under MIT license (see LICENSE file for details)
 
-using Newtonsoft.Json;
+using System.Text.Json.Serialization;
 using Silverback.Messaging.Messages;
 
 namespace Silverback.Tests.Integration.E2E.TestTypes.Messages